chore: better file dir
