+++ constants.ts
@@ -1,0 +1,5 @@
+export const FAVORITE_STORAGE_KEY = {
+  cafe: 'favoriteCafeList',
+  mart: 'favoriteMartList',
+  shuttle: 'favoriteShuttleList',
+};
