+++ helpers/slack.ts
@@ -1,0 +1,5 @@
+import {IncomingWebhook} from '@slack/webhook';
+
+export const slack = new IncomingWebhook(
+  'https://hooks.slack.com/services/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',
+);
