+++ screens/Mart.tsx
@@ -3,7 +3,7 @@ import React, {useEffect, useState} from 'react';
 import {compareAsc, getDay, parse as parseTime} from 'date-fns';
 import Grid from '../components/Grid';
 import {ParamListBase} from '@react-navigation/native';
-import {FAVORITE_STORAGE_KEY} from '../App';
+import {FAVORITE_STORAGE_KEY} from '../constants';
 
 type Props = BottomTabScreenProps<ParamListBase, '편의점'> & {
   marts: Mart[];
